fix minor error & add deployment doc

+++ database/migrations/2017_10_02_182407_create_participants_table.php
@@ -19,7 +19,7 @@ class CreateParticipantsTable extends Migration
             $table->string('lastname');
             $table->integer('age');
             $table->string('adress');
-            $table->tinyInteger('housenumber');
+            $table->integer('housenumber');
             $table->string('municipality');
             $table->integer('postalcode');
             $table->string('email', 100)->unique();
